Segregated Renderer into subfiles and it's own directory Should be easier to keep track of it
