+++ src/json/debug.json
@@ -7,7 +7,7 @@
             "#SSINIT_input": "Input manager initialized",
             "#SSINIT_json": "Json manager initialized"
         },
-        "#SYS_processSpeed": "Finished processing in:"
+        "#SYS_processSpeed": "Previous tick finished processing in:"
     },
     ".DEBUG_input":{
         "#INPUT_whatKey": "Key pressed:"
